+++ engine/render/RenderSnapshot.h
@@ -14,7 +14,7 @@ struct EntityRenderEntry {
     uint32_t entityIdx{0};                                     // EntityId::index — pool slot number
     uint32_t entityGen{0};                                     // EntityId::generation — stale-handle discriminator
     uint32_t typeIndex{0};                                     // EntityState::typeIndex — index into EntityTypeRegistry
-    glm::vec3 position{};                                      // world position (m)
+    glm::dvec3 position{};                                     // world position (m) — double for planet-scale precision
     glm::quat orientation = glm::quat(1.0f, 0.0f, 0.0f, 0.0f); // world orientation (identity)
     glm::vec3 velocity{};                                      // world velocity (m/s) — used for sub-tick extrapolation
     uint8_t damageLevel{0};                                    // cast from DamageLevel; 0 = Intact
